完善操作权限，只有面试的老师才有资格修改或删除对应学生的信息，其他咨询没有资格。

+++ oa/src/main/java/com/hotmarzz/oa/buzz/impl/StudentBuzzImpl.java
@@ -6,6 +6,9 @@ import java.util.List;
 import javax.servlet.http.HttpSession;
 
 
+
+
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +18,8 @@ import com.hotmarzz.basic.dao.BaseQuery;
 import com.hotmarzz.oa.buzz.StudentBuzz;
 import com.hotmarzz.oa.controller.EmpController;
 import com.hotmarzz.oa.dao.StudentDao;
+import com.hotmarzz.oa.exception.CanNotDeleteStuException;
+import com.hotmarzz.oa.exception.CanNotUpdateStuException;
 import com.hotmarzz.oa.exception.StudentRepeatException;
 import com.hotmarzz.oa.pojo.Emp;
 import com.hotmarzz.oa.pojo.Student;
@@ -25,7 +30,6 @@ public class StudentBuzzImpl implements StudentBuzz {
 	
 	private Logger logger = LoggerFactory.getLogger(StudentBuzzImpl.class);
 
-	
 	private StudentDao stuDao;
 
 	public StudentDao getStuDao() {
@@ -56,6 +60,21 @@ public class StudentBuzzImpl implements StudentBuzz {
 
 	@Override
 	public void update(Student stu) throws Exception {
+		Student stuForUpdate=(Student) session.getAttribute(SessionUtils.FOR_UPDATE);
+		String stuName1=stu.getStuName();
+		int sex1=stu.getSex();
+		int age1=stu.getStuAge();
+		
+		String stuName2=stuForUpdate.getStuName();
+		int sex2=stuForUpdate.getSex();
+		int age2=stuForUpdate.getStuAge();
+		//判断是否为同一个学生数据
+		if(!(stuName1.equals(stuName2) && sex1==sex2 && age1==age2)){
+			String lockUserName = checkStuRepeat(stu);
+			if(lockUserName!=null){
+				throw new StudentRepeatException(lockUserName);
+			}
+		}
 		stuDao.update(stu);
 	}
 
@@ -65,6 +84,21 @@ public class StudentBuzzImpl implements StudentBuzz {
 	@Override
 	public Student getById(Long id) throws Exception {
 		return stuDao.getById(id);
+	}
+	//检测是否能修改专用
+	@Override
+	public void getById1(Long id) throws Exception {
+		Emp e=(Emp)session.getAttribute(SessionUtils.LOGIN_EMP_KEY);
+		String loginName=null;
+		if(e!=null){
+			loginName=e.getEmpName();
+		}
+		
+		String locker=stuDao.getById(id).getLockUser();
+		
+		if(loginName==null || !loginName.equals(locker)){
+			throw new CanNotUpdateStuException();
+		}
 	}
 
 	/**
@@ -84,6 +118,17 @@ public class StudentBuzzImpl implements StudentBuzz {
 	 */
 	@Override
 	public void delete(Long stuId) throws Exception {
+		Emp e=(Emp)session.getAttribute(SessionUtils.LOGIN_EMP_KEY);
+		String loginName=null;
+		if(e!=null){
+			loginName=e.getEmpName();
+		}
+		
+		String locker=stuDao.getById(stuId).getLockUser();
+		
+		if(loginName==null || !loginName.equals(locker)){
+			throw new CanNotDeleteStuException();
+		}
 		stuDao.delete(stuId);
 	}
